Setup client project
