+++ client/Lykke.Snow.Notifications.Client/Model/DeviceRegistrationErrorCodeContract.cs
@@ -1,0 +1,29 @@
+namespace Lykke.Snow.Notifications.Client.Model
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public enum DeviceRegistrationErrorCodeContract
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        None,
+        /// <summary>
+        /// 
+        /// </summary>
+        DoesNotExist,
+        /// <summary>
+        /// 
+        /// </summary>
+        AlreadyRegistered,
+        /// <summary>
+        /// 
+        /// </summary>
+        AccountIdNotValid,
+        /// <summary>
+        /// 
+        /// </summary>
+        DeviceTokenNotValid
+    }
+}
