--- contract/Lykke.Snow.Notifications.Contract/Model/Contracts/DeviceRegistrationErrorCodeContract.cs
@@ -1,11 +1,0 @@
-namespace Lykke.Snow.Notifications.Contracts.Model.Contracts
-{
-    public enum DeviceRegistrationErrorCodeContract
-    {
-        None,
-        DoesNotExist,
-        AlreadyRegistered,
-        AccountIdNotValid,
-        DeviceTokenNotValid
-    }
-}
