+++ src/Lykke.Snow.Notifications/Controllers/DeviceRegistrationController.cs
@@ -3,8 +3,8 @@ using System.Threading.Tasks;
 using AutoMapper;
 using Lykke.Snow.Common.Model;
 using Lykke.Snow.Contracts.Responses;
-using Lykke.Snow.Notifications.Contracts.Model.Contracts;
-using Lykke.Snow.Notifications.Contracts.Model.Requests;
+using Lykke.Snow.Notifications.Client.Model;
+using Lykke.Snow.Notifications.Client.Model.Requests;
 using Lykke.Snow.Notifications.Domain.Enums;
 using Lykke.Snow.Notifications.Domain.Model;
 using Lykke.Snow.Notifications.Domain.Services;
@@ -40,7 +40,7 @@ namespace Lykke.Snow.Notifications.Controllers
 
         [HttpPost("unregister")]
         [ProducesResponseType(typeof(ErrorCodeResponse<DeviceRegistrationErrorCodeContract>), (int) HttpStatusCode.OK)]
-        public async Task<ErrorCodeResponse<DeviceRegistrationErrorCodeContract>> DeregisterDevice(UnregisterDeviceRequest request)
+        public async Task<ErrorCodeResponse<DeviceRegistrationErrorCodeContract>> UnregisterDevice(UnregisterDeviceRequest request)
         {
             var response = new ErrorCodeResponse<DeviceRegistrationErrorCodeContract>();
