+++ src/Lykke.Snow.Notifications/Lykke.Snow.Notifications.csproj
@@ -17,7 +17,7 @@
     <ProjectReference Include="..\Lykke.Snow.Notifications.Domain\Lykke.Snow.Notifications.Domain.csproj" />
     <ProjectReference Include="..\Lykke.Snow.Notifications.DomainServices\Lykke.Snow.Notifications.DomainServices.csproj" />
     <ProjectReference Include="..\Lykke.Snow.Notifications.SqlRepositories\Lykke.Snow.Notifications.SqlRepositories.csproj" />
-    <ProjectReference Include="..\..\contract\Lykke.Snow.Notifications.Contract\Lykke.Snow.Notifications.Contract.csproj" />
+    <ProjectReference Include="..\..\client\Lykke.Snow.Notifications.Client\Lykke.Snow.Notifications.Client.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="8.0.0" />
